[#69] Renamed BinaryFile to BinaryData
Does not write to file so made more sense.
