+++ SuperGameEngine/FatedQuest.Libraries/BinaryOperations/BinaryFile/StandardBinaryData.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "../../StandardCLibrary/LibraryIncludes.h"
-#include "BinaryFile.h"
+#include "BinaryData.h"
 #include "BinaryString.h"
 #include "BinaryInt.h"
 #include <map>
@@ -14,18 +14,18 @@ namespace BinaryOperations
     /// A binary file using the StandardCLibrary.
     /// This is the type to use for SuperGameEngine.
     /// </summary>
-    class StandardBinaryFile : public BinaryFile
+    class StandardBinaryData : public BinaryData
     {
     public:
-        StandardBinaryFile();
-        virtual ~StandardBinaryFile();
+        StandardBinaryData();
+        virtual ~StandardBinaryData();
 
         /// <summary>
         /// Add an int to the file.
         /// This means this int would be considered as part of the reading
         /// and writing process.
         /// </summary>
-        /// <param name="key">What to call this int.
+        /// <param name="key">What to call this int. </param>
         /// This must be a unique name for this int.</param>
         virtual void AddInt(std::string key) override;
 
@@ -44,6 +44,31 @@ namespace BinaryOperations
         /// <param name="defaultValue">Default value to use if not set. </param>
         /// <returns>Value of the int or default if not found. </returns>
         virtual int TryGetInt(std::string key, int defaultValue) override;
+
+        /// <summary>
+        /// Add an string to the file.
+        /// This means this int would be considered as part of the reading
+        /// and writing process.
+        /// </summary>
+        /// <param name="key">What to call this string. </parm>
+        /// This must be a unique name for this int.</param>
+        virtual void AddString(std::string key) override;
+
+        /// <summary>
+        /// Set the value of an string.
+        /// </summary>
+        /// <param name="key">Unique name for the string. </param>
+        /// <param name="value">New value to set. </param>
+        virtual void SetString(std::string key, std::string value) override;
+
+        /// <summary>
+        /// Try to get the value of the string.
+        /// Returns default if not set.
+        /// </summary>
+        /// <param name="key">Unique name for the string. </param>
+        /// <param name="defaultValue">Default value to use if not set. </param>
+        /// <returns>Value of the string or default if not found. </returns>
+        virtual std::string TryGetString(std::string key, std::string defaultValue) override;
 
         /// <summary>
         /// Export the data in a binary form.
@@ -73,6 +98,21 @@ namespace BinaryOperations
         /// Holds the Int values for the Binary Operations.
         /// </summary>
         std::map<std::string, IntValue> m_intValues;
+
+        /// <summary>
+        /// Indicates if an string value is set or not.
+        /// </summary>
+        struct StringValue
+        {
+        public:
+            std::string Value = std::string();
+            bool IsSet = false;
+        };
+
+        /// <summary>
+        /// Holds the String values for the Binary Operations.
+        /// </summary>
+        std::map<std::string, StringValue> m_stringValues;
 
         /// <summary>
         /// Operations for binary strings.
@@ -115,5 +155,13 @@ namespace BinaryOperations
         /// <param name="stringStream">String stream seeked to the int keys. </param>
         /// <returns>Current seeked string. </returns>
         std::string ProcessIntKeys(std::shared_ptr<FStringStream> stringStream);
+
+        /// <summary>
+        /// Take the binary data and load the string key from it.
+        /// Must start at the point the int keys begin.
+        /// </summary>
+        /// <param name="stringStream">String stream seeked to the string keys. </param>
+        /// <returns>Current seeked string. </returns>
+        std::string ProcessStringKeys(std::shared_ptr<FStringStream> stringStream);
     };
 }
